refactor(main): main プロセス専用になった ApmJson・shortcut を src/main へ移動
Phase 3 の main 化で renderer からの利用者がいなくなったため、
「src/lib = renderer 系」の区分に合わせて配置を修正する。
Config は renderer と main の両方が electron-store のディスク同期を
前提に共有するため src/lib に残す。
