+++ src/main/index.ts
@@ -4,9 +4,9 @@ import log from 'electron-log/main';
 import Store from 'electron-store';
 import 'source-map-support/register';
 import { getConfig } from '../lib/Config';
-import * as shortcut from '../lib/shortcut';
 import { registerIpcHandlers } from './ipcHandlers';
 import { ensureAutoUpdateDefault } from './services/appUpdate';
+import * as shortcut from './shortcut';
 import { launch } from './windows';
 
 log.initialize();
