+++ src/main/services/core.ts
@@ -3,13 +3,13 @@ import { app, type BrowserWindow, dialog } from 'electron';
 import log from 'electron-log/main';
 import { existsSync, readJson } from 'fs-extra';
 import path from 'node:path';
-import ApmJson from '../../lib/ApmJson';
 import type Config from '../../lib/Config';
-import { addAviUtlShortcut, removeAviUtlShortcut } from '../../lib/shortcut';
 import { installedVersionText } from '../../shared/coreVersionText';
 import { install, verifyFilesByCount } from '../../shared/install';
 import { checkIntegrity, verifyFile } from '../../shared/integrity';
 import unzip from '../../shared/unzip';
+import ApmJson from '../ApmJson';
+import { addAviUtlShortcut, removeAviUtlShortcut } from '../shortcut';
 import { downloadFile } from './download';
 import { migrationByFolder } from './migration';
 import { getCoreDataUrl, getInfo, updateInfo } from './modList';
