+++ src/main/shortcut.ts
